+++ src/Resources/contao/languages/de/default.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['TL_LANG']['CTE']['spielerregister_person'] = array('Person aus dem Spielerregister', 'Person aus dem Spielerregister anzeigen');