+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,0 +1,5 @@
+<?php 
+
+$GLOBALS['TL_LANG']['tl_content']['spielerregister_legend'] = 'Einstellungen';
+$GLOBALS['TL_LANG']['tl_content']['spielerregister_id'] = array('Person auswählen', 'Bitte wählen Sie die anzuzeigende Person aus. Vor- und Nachname der Person werden in die Überschrift eingebaut, falls dort ein Platzhalter %s steht.');
+//$GLOBALS['TL_LANG']['tl_content']['spielerregister_overwriteHeadline'] = array('Überschrift überschreiben/ergänzen', 'Überschreibt die Überschrift mit Vor- und Nachname. Steht in der Überschrift ein Platzhalter %s, wird dort der Vor- und Nachname eingesetzt.');